chore: Update RR scripts for CJS
