--- contracts/scripts/v0.1/request-response/read-data.ts
@@ -1,14 +1,0 @@
-import { ethers } from 'hardhat'
-
-async function main() {
-  const userContract = await ethers.getContract('RequestResponseConsumerMock')
-  console.log('RequestResponseConsumerMock', userContract.address)
-
-  const response = await userContract.sResponse()
-  console.log(`Response ${response}`)
-}
-
-main().catch((error) => {
-  console.error(error)
-  process.exitCode = 1
-})
