+++ contracts/scripts/v0.1/request-response/request-data-direct.cjs
@@ -1,5 +1,5 @@
-import { ethers } from 'hardhat'
-import hre from 'hardhat'
+const { ethers } = require('hardhat')
+const hre = require('hardhat')
 
 async function main() {
   const requestResponseConsumerMock = await ethers.getContract('RequestResponseConsumerMock')
@@ -12,9 +12,16 @@ async function main() {
   )
 
   const callbackGasLimit = 500_000
-  await requestResponseConsumerSigner.requestDataDirectPayment(callbackGasLimit, {
-    value: ethers.utils.parseEther('1.0')
-  })
+  const numSubmission = 1
+  const refundRecipient = requestResponseConsumerSigner.address
+  await requestResponseConsumerSigner.requestDataDirectPaymentUint128(
+    callbackGasLimit,
+    numSubmission,
+    refundRecipient,
+    {
+      value: ethers.utils.parseEther('1.0')
+    }
+  )
 }
 
 main().catch((error) => {
