+++ contracts/scripts/v0.1/vrf/README.md
@@ -6,17 +6,17 @@ To deploy the smart contracts, run `npx hardhat deploy --network localhost`.
 ## Request VRF
 
 ```shell
-npx hardhat run request-vrf.ts --network localhost
+npx hardhat run request-vrf.cjs --network localhost
 ```
 
 ## Request VRF with direct payment
 
 ```shell
-npx hardhat run request-vrf-direct.ts --network localhost
+npx hardhat run request-vrf-direct.cjs --network localhost
 ```
 
 ## Read VRF response
 
 ```shell
-npx hardhat run read-vrf.ts --network localhost
+npx hardhat run read-vrf.cjs --network localhost
 ```
